% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_similarities.R
\name{average_similarities}
\alias{average_similarities}
\title{Average Similarities}
\usage{
average_similarities(duplicates_df, similarities, plot = TRUE)
}
\arguments{
\item{duplicates_df}{data frame containing duplicate entries of trip comparisons}

\item{similarities}{data frame with all pairwise trip comparisons}

\item{plot}{if TRUE, print plot to compare distributions of duplicated v. averaged similarities}
}
\value{
similarity data frame without duplicates, and new averaged similarity values
}
\description{
Some pairwise trip comparisons will be duplicated if similarities
are calculated by gear group. This function averages the similarities
from duplicated comparisons, and replaces all duplicates in the full
data frame of pairwise trip similarities. For script 01, section 4.2
}
\examples{
similarities <- average_similarities(duplicates_df=duplicate_rows, similarities = similarities, plot=TRUE)
}
