% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stats_functions.R
\name{beta_eff}
\alias{beta_eff}
\title{Functions for Network Statistics}
\usage{
beta_eff(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
numeric value
}
\description{
These functions calculate the following network statistics: (1) beta efficiency,
network centralization (unweighted), (3) network centralization (weighted). They
are called within the net_stats function. Network #' centralization code is
based on equations from Cinner & Bodin (2010), doi: 10.1371/journal.pone.0011999
}
\examples{

}
