% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_length.R
\name{calc_length}
\alias{calc_length}
\title{Calculate Vessel Lengths from Registration Data}
\usage{
calc_length(permits, vesseldat, lengthdat, summarydat, index)
}
\arguments{
\item{permits}{vessel registration data frame}

\item{vesseldat}{vessel IDs for fishing vessels active in a given year}

\item{lengthdat}{vessel length from registration data frame}

\item{summarydat}{a data frame with contains: VESSEL_NUM, AGENCY_CODE, REGISTRATION_YEAR, n_lengths, n_unique, max_length, min_length, median2yr}

\item{index}{index, representing the position of the relevant vessel id in the vesseldat data frame}
}
\value{
the final vessel length, number of total / unique lengths in the registration data, and the length calculation used.
}
\description{
Get vessel length, based on either 3 or 5 years of vessel
registration data to account for recording error (which is fairly
frequent in the PacFIN registration data set. The function can
write out warning messages as it runs about which vessels are
missing data, and which vessel lengths may be inaccurate because
they had to be pulled from archived registration data (4+ years ago).
This is originally from the script
Report_Vessel_Length_processedtix_v5_functions.R and is also used in
the VMS-Fish Ticket matching pipeline (see github.com/owenrliu)
For: \code{script_00}
}
\examples{
tmp_vessel_length_info <- calc_length(permits=permits, vesseldat = year_vessels, lengthdat = pthin_length_filter, summarydat = pthin_sumstats, index = i)
}
