% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_network_collapse_confidential.R
\name{collapse_confidential_early}
\alias{collapse_confidential_early}
\title{Collapse Confidential Fisheries, Early Season Directed Network}
\usage{
collapse_confidential_early(
  A,
  vpf_2014,
  vpf_2015,
  no_drop = c("DCRB_POT", "other_port", "no_fishing")
)
}
\arguments{
\item{A}{confidential adjacency matrix from \code{gen_adj_matrix_early} function}

\item{vpf_2014}{data frame with 2014-15 vessel counts per fishery from \code{gen_adj_matrix_early} function}

\item{vpf_2015}{data frame with 2015-16 vessel counts per fishery from \code{gen_adj_matrix_early} function}

\item{no_drop}{specify which nodes \emph{not} to collapse (if confidential will just be dropped)}
}
\value{
a list with (1) A: adjacency matrix with non-confidential data to create directed network, (2 & 3) vpf_2014 & vpf_2015: vessels participating in each fishery / aggregate in 2014-15 and 2015-16
}
\description{
Collapse confidential fisheries (< 3 vessels) in the early season directed
networks. Confidential Pot / Hook and Line fisheries are collapsed to an
'other, pot/hkl' fishery, unless a self-loop has more than three vessels.
If there are no confidential pot / hook and line fisheries, or if aggregating
those fisheries still results in participation of < 3 vessels, confidential
fisheries are collapsed into a more generic 'other' fishery. For Script 09a.
}
\examples{
new_A_info <- collapse_confidential_early(A=A,vpf_2014=vpf_2014,vpf_2015=vpf_2015)
}
