% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_confidential_matrix_rows.R
\name{collapse_rowwise}
\alias{collapse_rowwise}
\title{Collapse Individual Vessel Participation in Confidential Fisheries, Late Season Directed Network}
\usage{
collapse_rowwise(x, prioritize = met_mode)
}
\arguments{
\item{x}{a row of participation codes (1,-1,2,0) representing a single vessel}

\item{prioritize}{if a vessel has multiple participation codes for a single collapsed node, and there is no true mode, which code should be returned?}
}
\value{
a single participation code for a collapsed network node
}
\description{
Collapses a row (one vessel) of pre-defined confidential fisheries in the
late season directed networks.
Called by the \code{collapse_confidential_late} function.
}
\examples{
collapsed_values <- unlist(apply(vesselmat_con[,pl_metiers], 1, collapse_rowwise, default_other))
}
