% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_distance_matrix.R
\name{convertDist}
\alias{convertDist}
\title{Convert Distance Matrix}
\usage{
convertDist(inDist)
}
\arguments{
\item{inDist}{matrix output from of pairwise distances from parDist function}
}
\value{
data frame of distance values for each pairwise comparison in the matrix
}
\description{
Convert a matrix of distance values into a data frame. For script 01, section 2.3
}
\examples{
hellinger_df_list <- hellinger_dist_list \%>\% map(~convertDist(.x))
}
