% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_directed_network_early_supportfx_base.R
\name{directed_early_base}
\alias{directed_early_base}
\title{Basic Graph, Early Season Directed Network}
\usage{
directed_early_base(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
a ggraph object
}
\description{
Base plotting function to graph a directed network.
Called from within \code{graph_directed_early.R}
}
\examples{
plot_directed_early(g=mygraph)
}
