% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_directed_network_early_supportfx_coords.R
\name{directed_early_coords}
\alias{directed_early_coords}
\title{Fixed Coordinates Graph, Early Season Directed Network}
\usage{
directed_early_coords(g, dist_y = 1.2)
}
\arguments{
\item{g}{igraph object}

\item{dist_y}{set the vertical distance between network nodes}
}
\value{
a ggraph object
}
\description{
Graph a directed network with nodes at manually fixed x and y coordinates.
Called from within \code{graph_directed_early.R}
}
\examples{
plot_directed_early(g=mygraph, dist_y=1)
}
