% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_similarity.R
\name{dist2sim}
\alias{dist2sim}
\title{Convert Distance to Similarities}
\usage{
dist2sim(x, col, max.type = "individual", total.max = 0)
}
\arguments{
\item{x}{data frame distance values}

\item{col}{column number with distance values}

\item{max.type}{If "individual", the maximum will be calculated from the given matrix. If "manual", a specified value will be applied as the maximum}

\item{total.max}{If max.type = "manual", provide the maximum value to be subtracted from each distance}
}
\value{
data frame with pairwise similarity values
}
\description{
The Hellinger distance is used to quantify difference between trips,
but the infomap function requires similarity values. This function will
obtain a similarity score by subtracting the maximum value of D from
each pairwise comparison's distance value. For script 01, section 3.1
}
\examples{
similarities_df_list <- hellinger_df_list \%>\% map(~dist2sim(.x,3, max.type = "manual", total.max = dist_max))
}
