% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_adjacency_matrix.R
\name{fill_adjacency_matrix}
\alias{fill_adjacency_matrix}
\title{Fill an adjacency matrix}
\usage{
fill_adjacency_matrix(vessel_matrix, self = TRUE)
}
\arguments{
\item{vessel_matrix}{a matrix with vessels in rows, metiers in columns, and participation coded as (0,1,-1,2)}

\item{self}{record continued participation in a fishery? (represented by the diagonal in the adjacency matrix, and a self-loop in the network graph)}
}
\value{
a fisheries x fisheries adjacency matrix, with cell entries as vessel counts
}
\description{
Fill a fisheries x fisheries adjacency matrix, the foundation
of the network graphs.
Called by the \code{collapse_confidential_late} function.
}
\examples{
collapsed_values <- unlist(apply(vesselmat_con[,pl_metiers], 1, collapse_rowwise, default_other))
}
