% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directed_network_matrix_early.R
\name{gen_adj_matrix_early}
\alias{gen_adj_matrix_early}
\title{Generate Directed Network, Early Season}
\usage{
gen_adj_matrix_early(
  fishtix,
  p,
  self = FALSE,
  write = FALSE,
  outdir = NULL,
  size = NA,
  noncrab_fishtix = NULL
)
}
\arguments{
\item{p}{specify a port group}

\item{self}{calculate the diagonal of the matrix}

\item{write}{write out the adjacency matrix to a .csv file}

\item{outdir}{if write is TRUE, the directory to save the adjacency matrix}

\item{size}{vessel size category \link{large / small / NA}. used for file name if write=TRUE}

\item{noncrab_fishtix}{fish ticket data frame from non-crab ports (used for looking for vessel movement)}

\item{tickets}{fish tickets data frame}
}
\value{
a list with (1) A: adjacency matrix to create directed network, (2) n_dcrb_vessels: total number of vessels, (3 & 4) vpf_2014 & vpf_2015: vessels participating in each fishery in 2014-15 and 2015-16
}
\description{
Create a directed network to show flow of vessels between the 2015
Dungeness crab fishery and 2016 alternatives during the early season.
Original function name: vessel_adjacency_matrix_dcrb. For script 09a
}
\examples{
A_info <- gen_adj_matrix_early(fishtix_lg,p=p,self=TRUE,write=TRUE,size="large",noncrab_fishtix=noncrab_tix)
}
