% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_catch_matrix_reruns.R
\name{gen_catch_matrix_rerun}
\alias{gen_catch_matrix_rerun}
\title{Generate Catch Matrix, InfoMap R. 2}
\usage{
gen_catch_matrix_rerun(
  mydata,
  multigear = FALSE,
  NAto0 = TRUE,
  adjust.revenue = TRUE
)
}
\arguments{
\item{mydata}{catch matrix}

\item{adjust.revenue}{If TRUE, use the adjusted revenue column; If FALSE, use raw recorded revenue}

\item{metric}{values for the matrix are exvessel revenue ('revenue') or landed pounds ('lbs')}

\item{gear}{If TRUE, columns will represent all species / gear group combinations; if FALSE, columns will represent all species}

\item{Nato0}{If TRUE, replace NA values with 0 (revenue or lbs) when matrix is spread}
}
\value{
a list with (1) catch matrix populated by revenue or landed pounds, (2) a melted data frame from catch matrix
}
\description{
From the fish ticket data, create a catch matrix (rows= trips,
columns = identifiers plus all species/gear combos, values = revenue or pounds).
Script 04, Section 2
}
\examples{
revmat <- gen_catch_matrix_rerun(tickets = portdat, multigear=TRUE, NAto0 = TRUE)
}
