% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_matrix.R
\name{generate_catch_matrix_knn}
\alias{generate_catch_matrix_knn}
\title{Generate Catch Matrix for KNN}
\usage{
generate_catch_matrix_knn(
  tickets,
  multigear = FALSE,
  NAto0 = TRUE,
  adjust.revenue = TRUE
)
}
\arguments{
\item{tickets}{fish ticket data without metiers}

\item{multigear}{If TRUE, builds matrix using species/gear combinations; if FALSE, just uses species}

\item{NAto0}{If TRUE, converts NAs to 0s in the catch matrix}

\item{adjust.revenue}{If TRUE, uses the adjusted revenue in the catch matrix; if FALSE, raw reported exvessel revenue}
}
\value{
catch matrix as data frame
}
\description{
Create a catch matrix (rows=trips, columns=species/gear combos, values = revenue)
for fish tickets that have not been assigned metiers with infomap. This will be
called from within the assign_metiers_knn function. It is very similar to the
generate_catch_matrix function used in script 01, but is specialized for the KNN code.
}
\examples{

}
