% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_edges.R
\name{generate_network_edges}
\alias{generate_network_edges}
\title{Generate Network Edges Data Frame}
\usage{
generate_network_edges(similarities, vertices)
}
\arguments{
\item{similarities}{data frame with pairwise trip similarity scores}

\item{vertices}{trip IDs data frame used as input for infomap algorithm}
}
\value{
a data frame to build network edges in igraph
}
\description{
Generate a data frame that matches pairs of nodes to the weight of the edge between
them, quantified by similarity value. For script 01, section 5.1
}
\examples{
edges_df <- generate_network_edges(similarities = similarities, vertices = vertices_df)
}
