% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths_2yr_median.R
\name{get2yrmedian}
\alias{get2yrmedian}
\title{Calculate Median of Vessel Lengths}
\usage{
get2yrmedian(x, years)
}
\arguments{
\item{years}{vector of corresponding registration years}

\item{v}{vector of vessel lengths for individual vessel}
}
\value{
median (numeric)
}
\description{
Calculates the median from a vector of vessel lengths.
}
\examples{
getmode(v=c(30,30,3))
}
