% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths_5yr_registration.R
\name{get_historic_length}
\alias{get_historic_length}
\title{Calculate Length from 5 years of Registration Data}
\usage{
get_historic_length(tmp_vessel_info)
}
\arguments{
\item{tmp_vessel_info}{a data frame with contains: VESSEL_NUM, AGENCY_CODE, REGISTRATION_YEAR, n_lengths, n_unique, max_length, min_length, median2yr}
}
\value{
the final vessel length, and the type of calculation used to arrive at that length
}
\description{
Calculate the vessel length based on 5 years of
vessel registration data. For vessels which do not
have adequate registration data for the most recent
3 years. Called by \code{get_vessel_length.R} in \code{script_00}
}
\examples{
olddat_output <- get_historic_length(tmp_vessel_info=old_vessel_info)
}
