% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_glm.R
\name{get_sc_coef}
\alias{get_sc_coef}
\title{Standardize GLM}
\usage{
get_sc_coef(m, name, season)
}
\arguments{
\item{m}{the unstandardized model from the "glm" function}

\item{name}{the network metric being modeled (as a character string)}

\item{season}{whether the input data are for the early or late season \link{"Early Season" / "Late Season"}}
}
\value{
a data frame with the standardized model's coefficients
}
\description{
Standardize the input variables for a GLM
}
\examples{
ed_early.glm <- glm(ed ~ D*R + N + pcgroup, data = edata, family = quasibinomial('logit'))
ed_early.sglm <- get_sc_coef(ed.glm, name="Edge Density",season="Early Season")
}
