% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{getmode}
\alias{getmode}
\title{Calculate Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector of numbers or integers}
}
\value{
mode (numeric)
}
\description{
Simply calculates the mode from a vector of numbers.
The second version excludes zeros when calculating the mode.
}
\examples{
getmode(v=c(30,30,3))
}
