% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_metiers.R
\name{identify_metiers}
\alias{identify_metiers}
\title{Identify Metiers with InfoMap}
\usage{
identify_metiers(vertices, edges, trials = 100)
}
\arguments{
\item{vertices}{data frame of vertices / nodes (individual trips)}

\item{edges}{data frame of edges (similarity values for pairwise trip comparisons)}

\item{trials}{The number of attempts to partition the network for infomap algorithm (>1)}
}
\value{
an object from the igraph communities class
}
\description{
Create a graph object with the given vertices and edges, and then
use the InfoMap function in the igraph package to identify metiers
with the infomap algorithm. For script 01, section 6.1
}
\examples{
mycommunities <- identify_metiers(edges = edges_df, vertices = vertices_ext)
}
