% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assign.R
\name{knn_assign_metiers}
\alias{knn_assign_metiers}
\title{KNN Metier Assignment}
\usage{
knn_assign_metiers(
  train,
  key,
  test,
  k,
  adjust.revenue = TRUE,
  print.plot = TRUE
)
}
\arguments{
\item{train}{per-trip revenue data with metiers ('neighbors').}

\item{key}{A data frame with two columns: trip_id, and the identified metier from infomap}

\item{test}{per-trip revenue data without metiers}

\item{k}{The k value to use in the k-nearest neighbor algorithm}

\item{adjust.revenue}{If TRUE, uses the adjusted revenue in the catch matrix; if FALSE, raw reported exvessel revenue}

\item{print.plot}{If TRUE, plot membership to metiers}
}
\value{
data frame with assigned metiers
}
\description{
Assign new fish ticket data metiers based on the k-nearest neighbor algorithm (KNN),
which is conducted using customized R code in this function (instead of a hard-coded
R function). For use in script 03 section 1.2
}
\examples{
tmp_test_metiers <- assign_metiers_bygear(train = traindat, key = metier_key, test = tmp_tix, k = 2)
}
