% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_df}
\alias{merge_df}
\title{Merge to Similiarities Data Frame}
\usage{
merge_df(gear.list, sim.list, ids.list)
}
\arguments{
\item{gear.list}{character vector of all gear group types in the landings input data}

\item{sim.list}{list of data frames with pairwise trip similarities for each gear group}

\item{ids.list}{list of data frames with identifying information for each trip in the sim.list}
}
\value{
a data frame with all pairwise trip similarities + gear group column
}
\description{
For computational reasons, similarities between trips are calculated by gear group.
This function merges all gear group data frames to one. For script 01, section 4.1
}
\examples{
similarities <- merge_df(gear.list = gear_list_filtered, sim.list = similarities_df_list, ids.list = ids_bygear_list)
}
