% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier_key_reruns.R
\name{metier_key_rerun}
\alias{metier_key_rerun}
\title{Match Metiers to Trips}
\usage{
metier_key_rerun(revenue, vertices, name = TRUE)
}
\arguments{
\item{revenue}{data frame with revenue for each species/gear combination in each metier}

\item{vertices}{trips IDs data frame used as input for infomap algorithm}

\item{name}{if TRUE, generate a character name for each metier based on revenue per species/gear combination}
}
\value{
a data frame with trip IDs matched to a metier (numeric output from infomap) and a metier name (optional)
}
\description{
Match metiers (communities) to trip IDs, and automatically generate names
for each metier (community) identified by the infomap algorithm. For script 04, section 2
}
\examples{
trip_community_key <- metier_key_rerun(revenue = rev_by_community, vertices = vertices_ext)
}
