% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_participation_network.R
\name{participation_network_crabyear}
\alias{participation_network_crabyear}
\title{Generate Participation Network}
\usage{
participation_network_crabyear(
  tickets,
  pcid_choose = NA,
  year_choose = NA,
  filter,
  filter_subgraph,
  min_vessels = 3,
  min_contribution = 0.1
)
}
\arguments{
\item{tickets}{fish tickets data frame}

\item{pcid_choose}{specify a port group}

\item{year_choose}{Specify a crab year}

\item{filter}{use the \code{min_vessels} and \code{min_contribution} objects to filter the data}

\item{filter_subgraph}{a filtering option from the original function that was turned off for Fisher et al.}

\item{min_vessels}{the minimum number of vessels participating in a fishery for that fishery to be retained in the network}

\item{min_contribution}{the minimum contribution (as a proportion) to total exvessel revenue for a fishery to be retained for a given vessel}
}
\value{
non-confidential fisheries partition network as an igraph object
}
\description{
Create a fisheries participation network. Based on original function by
Emma Fuller; edits specified in comments.
}
\examples{
close_g <- participation_network_crabyear(close_dat, filter = TRUE, filter_subgraph = FALSE)
}
