% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_directed_network_early.R
\name{plot_directed_early}
\alias{plot_directed_early}
\title{Graph the Directed Network, Early Season}
\usage{
plot_directed_early(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
a ggraph object
}
\description{
Graph a directed network to show flow of vessels between the 2015
Dungeness crab fishery and 2016 alternatives during the early season.
Original function name: plot_graph. For script Fig3.Rmd
}
\examples{
plot_directed_early(g=mygraph)
}
