% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_directed_network_early_coords_image.R
\name{plot_directed_early_coords_image}
\alias{plot_directed_early_coords_image}
\title{Graph the Directed Network, Early Season}
\usage{
plot_directed_early_coords_image(
  g,
  dist_y = 1.2,
  image_object = img_file,
  image_size = 0.15,
  image_aspect = 1.2,
  image_offset = 0.15,
  image_transparent = FALSE
)
}
\arguments{
\item{g}{igraph object}

\item{dist_y}{specify the vertical distance between nodes, for when x_coords is TRUE}

\item{image_object}{a character string that provides the file name for the dungeness crab node icon, for when crab_image is TRUE}

\item{image_size}{the size of the dungeness crab node icon, for when crab_image is TRUE}

\item{image_aspect}{the aspect ratio (width:height) of the dungeness crab node icon, for when crab_image is TRUE}

\item{image_offset}{the distance to the left of the node to place the dungeness crab node icon, for when crab_image is TRUE}

\item{image_transparent}{the alpha value for the dungeness crab node icon, for when crab_image is TRUE}
}
\value{
a ggraph object
}
\description{
Graph a directed network to show flow of vessels between the 2015
Dungeness crab fishery and 2016 alternatives during the early season.
This is the function used for Fisher et al., although the plot_directed_early
function allows for several plotting methods. For script Fig3.Rmd
}
\examples{
plot_directed_early_coords_image(g=mygraph, image_object=here::here("data/input","crab_icon.png"))
}
