% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_directed_network_late.R
\name{plot_directed_late}
\alias{plot_directed_late}
\title{Graph the Directed Network, Late Season}
\usage{
plot_directed_late(g, dist_y = 1.8)
}
\arguments{
\item{g}{igraph object}

\item{dist_y}{set the vertical distance between network nodes (this is actually on the x axis due to coord_flip())}
}
\value{
a ggraph object
}
\description{
Graph a directed network to show flow of vessels between the 2015
Dungeness crab fishery and 2016 alternatives during the late season.
Original function name: plot_graph. For script FigS5.Rmd
}
\examples{
plot_directed_late(g=mygraph,dist_y=1.4)
}
