% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_network_nodes.R
\name{rename_vertices}
\alias{rename_vertices}
\title{Rename Network Nodes / Vertices}
\usage{
rename_vertices(g)
}
\arguments{
\item{g}{network as an igraph object}
}
\value{
igraph object with newly defined "common_name" attribute
}
\description{
User-friendly fishery names to replace the abbreviate species and gear codes
that define each node (vertex) in the network. Function is marked where
additional names can be hard-coded into the function. Theoretically, 'str_replace'
should work for all renaming, but I had some trouble with it as the pipeline
extended beyond 4-5 lines.
For scripts Fig3.Rmd and FigS5.Rmd
}
\examples{
vertex_cols <- vertex_color(g = early_igraph)
}
