% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_matrix.R
\name{split_matrix}
\alias{split_matrix}
\title{Split Catch Matrix by Gear Group}
\usage{
split_matrix(gear.list, catch_matrix)
}
\arguments{
\item{gear.list}{character vector of gear groups from the landings input data}

\item{catch_matrix}{catch matrix from 'generate_catch_matrix'}
}
\value{
(1) a list of catch matrices with only revenue / lbs split by gear group, (2) a list of matrices with identifying information for 1st outpu, (3) a revised character vector of gear types found in the catch matrix
}
\description{
For computational reasons, distances between trips are calculated by gear group.
his function will split the full catch matrix into a list of matrices, by gear group.
For script 01, section 3.1
}
\examples{
catchmats_bygear <- split_matrix(gear.list = gear_list, catch_matrix = inmat)
}
