% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_transparency.R
\name{transparent}
\alias{transparent}
\title{Image Transparency}
\usage{
transparent(img)
}
\arguments{
\item{img}{}
}
\value{
an image with adjusted transparency according to "0.8*a"
}
\description{
This function allows you to set image transparency
for the geom_image addition to ggplot. This function has
an alpha value hard-coded in. To change the alpha value,
replace the decimal in 'expression = X.X*a'.
}
\examples{
g + geom_image(aes(x=x, y=y, image=img_filename), image_fun=transparent, size=1, asp=1)
}
