% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_network_nodes.R
\name{vertex_color}
\alias{vertex_color}
\title{Color Network Nodes}
\usage{
vertex_color(mygraph, none = TRUE)
}
\arguments{
\item{g}{network as an igraph object}
}
\value{
character vector with colors for each node, ordered
}
\description{
Use the gear type and species ID to assign a color to
each node (vertex) in the network. Meant for both
directed and undirected network plotting functions.
}
\examples{
vertex_cols <- vertex_color(g = close_g); V(close_g)$colors <- vertex_col
}
